# Conflicts:
#	src/main/java/com/wediscussmovies/project/service/impl/ReplyServiceImpl.java
#	src/main/java/com/wediscussmovies/project/web/controller/MovieController.java
#	src/main/java/com/wediscussmovies/project/web/controller/ReplyController.java

+++ src/main/java/com/wediscussmovies/project/service/MovieService.java
@@ -1,6 +1,6 @@
 package com.wediscussmovies.project.service;
 
-import com.wediscussmovies.project.model.Movie;
+import com.wediscussmovies.project.model.*;
 
 import java.util.List;
 import java.util.Optional;
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface MovieService {
     public List<Movie> listAll();
     public List<Movie> searchByTitle(String title);
-    public Optional<Movie> findById(Long id);
+    public Optional<Movie> findById(Integer id);
     public Movie save(Movie movie);
-    public void deleteById(Long id);
+    public void deleteById(Integer id);
 }
